Add context menu options for new data structure
Changes the context menu options to match the new paradigm, so now there
is an add activation and remove activation to increment/decrement the
number of actions a unit has. By default a unit will have activations
equal to {actor}.data.data.activations, or 1 if that parameter is not
set.