Update theming; Bump compatible version
Adds configurable icon to settings allowing the GM to specify the css
icon classes to suit other systems. Moves static theming to a separate
stylesheet. Adds visual highlights to the left and right side of the
combat tracker using the same color coding as for the buttons. Theming
was tested with Ernie's Modern UI to ensure consistency.