Move logic to subclasses of Combat and CombatTracker
While robustness hasn't been improved, by moving the logic to
subclasses, the clarity of the changes being made is improved.
Additionally, this requires less boilerplate per changed method. Due to
this several additional methods have been wrapped or overridden in order
to expose fewer sharp edges from the shitty hard-coded assumptions of
the underlying classes.

The option to move activated units in the tracker has been reworked and
the logic moved into LancerCombatTracker to prevent updateCombat events
from not firing. With this rework, when units are moved in the turn
tracker after activating, the <li> tag is moved rather than the
position in the turn order. Additionally, units aren't moved if they are
active.

libWrapper is no longer used and has been removed. That was really short
lived, lol.