Add a custom settings dialog for icon settings
Moves the settings for the appearacnce of the icon to a custom dialog.
The dialog features a live preview of the icon to make tweaking the
apparacnce much quicker.

Icon size is now set by a direct config option. The old css classes have
been removed from the stylesheet. The new default size is larger. Sizing
is set in rem units and is thus relative to the root font size.

Add localization strings LANCERINITIATIVE.IconSettingsMenu
LANCERINITIATIVE.IconSize and LANCERINITIATIVE.IconPreview