Rework tracker sorting
Replaces the option to move units that have acted to the bottom of the
tracker with a more general sort option that moves the active unit to
the top of the tracker in addition to sorting units that have taken
their turn to the end of the tracker.

Removes the dummy combatant from the list of combatants passed to the
tracker. Because of this, there is no longer a need to hide it using
css.

Adds LANCERINITIATIVE.SortTracker and LANCERINITIATIVE.SortTrackerDesc
localization strings for the new option.