Set initative value on the dummy
Sets the dummy to have -1 for its initiative value so that the roll all
buttons do not attempt to roll initiative for the combatant. The roll
inititive function does not null check the token field before using it
and the position of the dummy at the start of the turn list meant that
it would bail out without rolling initiative for any combatants.