Move combat tracker modifications out of hooks.
Puts the modifications to the combat tracker into the _renderInner
method. This means that the the handlers can be set up in the
activateListeners method. Using jquery make the changes (slightly, and
only slightly) more readable, and is a much better option than using
handlebars absurdity to try to do this.