Change settings storage
Changes the settings storage to store the appearance data as an object.
Defines the defaults on the global CONFIG object. Includes migration
code so that customized settings aren't reset. This migration code will
be removed in the 0.5.x series. This is to make the eventual development
of this into a node package for inclusion in the Lancer system easier.