Initial port to typescript
Moves files and updates tooling for typescript.  Updates
foundry-pc-types to a personal fork. I may test with the upstream
version and possible rebase on that version in the future.

A Makefile or other build system is still needed and the github workflow
is will need to be updated to contain the build step.

Fixed missed settings change