Add a distinction between player and friendly combatants
Uses similar logic to token border to color the combatants. This also
increases the robustness of the coloring as now, if there is no token
for a combatant, the disposition will be pulled from the actors
prototype token. This allows correct coloring to be applied for theater
of the mind combat scenarios.