Properly skip defeated combatants
Uses the same logic as core foundry to determine which combatants are
defeated and set their activations to 0 on round change.

Uses foundry.utils.getProperty() to retrieve the activations value key.

Hide the roll initiative buttons at the top of the tracker when
initiative rolling is disabled.

Specify the parent when creating the dummy combatant to prevent foundry
from horribly exploding during the prepareData step for it.