Initial Star Wars FFG compatibility.
Use the base FFG combat class to create a mixin. As of right now, it's
not working, as the rollInitiative method from CombatFFG assumes that
there is always an active combatant and that turn is never null. I
should be able to submit a small patch to the system to fix that issue.